Modernize publicsuffix-data
- Obsolete callback-style usage
- Use native Promises
- Use more modern JS
- Upgrade to nyc
- Stop publishing extraneous files
- Update folder structure to something more typical
- Fix security audit failures
