+++ .gitignore
@@ -4,6 +4,7 @@ node_modules
 # Coverage
 lib-cov
 coverage
+.nyc_output
 
 # IDEs
 .idea
