+++ CHANGELOG.md
@@ -1,0 +1,7 @@
+# Changelog
+
+# 2.0.0
+
+Breaking:
+- Methods no-longer support callback-style usage; everything now uses Promises
+- Library now requires node@10
